+++ src/lib/scriptengine/rsscript.rs
@@ -76,7 +76,12 @@ fn create_rust_file(rust_script: &Vec<String>) -> String {
     create_persisted_script_file(rust_script, "rs")
 }
 
-fn run_file(file: &str, cli_arguments: &Vec<String>, provider: &ScriptRunner) -> bool {
+fn run_file(
+    file: &str,
+    runner_arguments: Option<Vec<String>>,
+    cli_arguments: &Vec<String>,
+    provider: &ScriptRunner,
+) -> bool {
     let (use_cargo, command) = match provider {
         ScriptRunner::RustScript => (false, "rust-script"),
         ScriptRunner::CargoScript => (true, "script"),
@@ -86,6 +91,9 @@ fn run_file(file: &str, cli_arguments: &Vec<String>, provider: &ScriptRunner) ->
     let mut args = vec![];
     if use_cargo {
         args.push(command.to_string());
+    }
+    if let Some(mut runner_args) = runner_arguments {
+        args.append(&mut runner_args);
     }
     args.push(file.to_string());
     let mut cli_args = cli_arguments.clone();
@@ -103,6 +111,7 @@ fn run_file(file: &str, cli_arguments: &Vec<String>, provider: &ScriptRunner) ->
 
 pub(crate) fn execute(
     rust_script: &Vec<String>,
+    runner_arguments: Option<Vec<String>>,
     cli_arguments: &Vec<String>,
     validate: bool,
 ) -> bool {
@@ -112,7 +121,7 @@ pub(crate) fn execute(
 
     let file = create_rust_file(rust_script);
 
-    let valid = run_file(&file, &cli_arguments, &provider);
+    let valid = run_file(&file, runner_arguments, &cli_arguments, &provider);
 
     if validate && !valid {
         error!("Unable to execute rust code.");